01-Web-App-Without-JS/requirements.txt
flask
flask-wtf
python-dotenv
auth0-python
